Improve notifications by tracking already notified movies
Implement functionality to save and load a file containing a list of movies that have been notified to users by the Discord bot (achieved through functions "write_notified_movies" and "read_notified_movies").

This file is checked against a list of upcoming movies recently pulled from TMBD, so that the bot sends notifications on new releases only. After checking for upcoming releases, the file is updated with the new releases.